Do not space after block comment characters
After block comment characters, remove the space. The space was intended
for if the comment characters were place right before text, to make it
easier to read. But now we put the comment characters at the beginning
of the line, since this more clearly states that all these lines of code
have been block commented, and they are not "real" comments so-to-speak.
Therefore, it's better just to replace the beginning of the line with
the comment characters, so that if one needs to remove the comment, they
can just delete the comment character, and not the comment + a space.