Merge pyimports with vimrc
Before, we had these files separated. Now, we've joined them because it
gets complicated since we are symlinking ~/.vimrc to vimrc.vim. In such
a case, we would also need to symlink pyimports.vim in ~. This workflow
may lead to a lot of symlinks in ~.

Later, it may be good to have these files, but have a bash script to set
up all the symlinks, and ensure they are all hidden.