Fix spacing in statusbar
Fixed some spacing in statusbar. The issue was only visible on Linux or
Windows systems, and it caused the colouring of the statusbar not to be
consistent with the spacing around the mode on the left-hand side of the
statusbar. The space before the mode string would be one colour, and the
spacing on the right hand side of the command mode string would be a
different colour. This has to do with the fact that the statusline is
not coloured on non-macOS systems, but is only coloured on macOS systems
based on the OS's appearance (light or dark).

The issue looked like:

1_BLACK_SPACE_==INSERT==_WHITE_SPACE_unix[gitcommit]...

where _BLACK_SPACE_ denotes a black space existed there and
_WHITE_SPACE_ donotes that a white space existed there. Now, they are
both the same colour, the colour of the statusbar, on non-macOS systems.

On macOS systems, the command mode string has the same colouring as the
statusbar. On non-macOS systems, they have different colours.