Change concealing, snipMate, and citations
Changed how concealing works, so it looks nicer. Also removed the
highlights on the concealed text.

Removed the ability to use snipMate snippets. UltiSnips work better and
I don't want them to conflict with each other.

Modified citations so they are now shown as a 📖 instead of in square
brackets. This saves a bunch of space.