Adjust spelling highlighting
Adjusted spelling highlighting so that it is easier to see. Now, when a
word is highlighted due to a spelling error, the text is changed to
white, instead of keeping its default colour.

Spelling mistakes are now underlined, rather than highlighted in red. In
code, there are a lot of spelling mistakes due to variable names, etc.,
and highlighting was too harsh, as it appeared everywhere. Underlining
looks much nicer.