Optimize startup
Changed up the code so the tabline and statusline are only themed with
their default themes if lightline is not being used. This also fixed a
bug where if vimrc was sourced after vim started up, then the lightline
theme would be replaced by the default theme for the statusline and
tabline.