Temporarily fix no tab completion in terminal
Before, the vim terminal mode tab completion did not work due to some
weird mapping which deal with switching buffer. Commented out that
mapping. Still not sure why that mapping affects tab completion in
the terminal as it has nothing to do with tabs. We need a permanent
fix for this once we deduce the reason this is happening.