Only use termguicolors on Darwin
Only use termguicolors on Darwin. On Linux, I use i3 + URxvt which does
not play super nicely with termguicolors set. So, on Linux,
termguicolours is not set and we instead use the ANSI colours for vim.
The monterey theme does deal with this case though, so we don't need an
additional theme for Linux.