Theme terminal with gui colors if termguicolors
Before, the terminal was themed with the gui colors regardless of
whether or not termguicolors was set. Now, gui colours are only used in
the terminal if termguicolors is set.

This was actually problematic, because based on the different terminals,
the GUI colours were slightly off from the terminal colours.

Removed termguicolors from vim in Ubuntu because the terminal keeps
having a black background.