Switch backtick and comma
Usually, comma moves the inverse of the previous f, F, t, or T command
used. But, I'm using the comma for jumping to marks, and so the
previously mentioned movements are overridden. So, since the comma has
taken over the meaning of the backtick, I've now set the backtick to
take over the meaning of the comma.